pkg: validate lockdir against local packages
This adds a type `Dune_pkg.Package_universe.t` made up of the set of
local packages in a project and a lockdir whose constructor validates
that the lockdir contains a valid solution for the dependencies of local
packages. As we add more pkg operations we'll need a way of talking
about the set of packages that makes up a project, both local and not.

Also this adds a command `dune pkg validate-lockdir` which validates
lockdirs against local packages in order to test this feature.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>