pkg: tidy up slang expansion
Replaces the polymorphic variant [`Concat] with the `deferred_concat`
type and thread it through the implementation of slang expansion. A list
of deferred concatenations is hopefully clearer to readers compared to
the `Value.t list list` it replaces.

Also use `option` rather than `list` internally to represent lists in
cases where it's statically known that their length will not exceed 1.
Codifying this fact allows us to remove a check for an error that was
impossible to encounter.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>