pkg: Set OCAMLFIND_DESTDIR for install actions
If a package uses ocamlfind to install its files then ocamlfind will
consult its config file to determine where in the filesystem the
installed files ought to go. In dune, each package is built in an
isolated sandbox, and so by default ocamlfind won't install files to the
correct location. The OCAMLFIND_DESTDIR environment variable can be used
to override ocamlfind's default behaviour for determining where to
install files.

This change sets the OCAMLFIND_DESTDIR variable to the "target/lib"
directory within a package's build sandbox when executing a package's
install command. Build commands are unaffected by this change.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>