Clarify some env handling for pkg rules
The logic for handling environments used when building packages is
complicated. This change breaks it up into smaller functions and adds
some comments explaining what each one does. A consequence of this
change is that the order in which the bin directories of dependencies
appear in the PATH variable while building a package is the same as
the order in which the dependencies appear in the package's
lockfile (previously the order was reversed).

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>