Clear console before running command in exec -w
This fixes an issue where the build progress stays on the console after
the build completes, and the first line of the output of the command
being exec'd is printed on teh same line as the build progress.

A side effect of this change is that some of the "Build Successful"
messages in tests get truncated.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>